size_t in mergepolys

+++ src/engine/world/octacube.cpp
@@ -650,7 +650,7 @@ void cube::mergepolys(int orient, const ivec &n, int offset, std::deque<poly> &p
     }
     std::unordered_set<plink> links(polys.size() <= 32 ? 128 : 1024);
     std::deque<const plink *> queue;
-    for(uint i = 0; i < polys.size(); i++)
+    for(size_t i = 0; i < polys.size(); i++)
     {
         const poly &p = polys[i];
         int prev = p.numverts-1;
